Linux port, reconstructed wam/data, and stage 1b event prediction
The `data/` gitignore pattern had no leading slash, so it matched a directory
named `data` at any depth and silently swallowed the whole `wam/data/` package.
It was never committed, and every entry point imports it. Anchored to `/data/`.

wam/data/ here is a RECONSTRUCTION, not the original. Window enumeration is
exact -- 423,316 candidates, matching docs/measurements.md to zero -- but the
instruction-labelling thresholds were back-solved from the published per-class
counts and land -2.6% overall. The per-class deviation table lives in
`contractor.FIT_QUALITY`; read any conclusion drawn from these labels next to it.

Stage 1b turns on the world head's event and flag prediction. Three of its four
outputs had been running forward every step without ever receiving a gradient,
because stage 1a loads only the image and action trees. Event prediction is the
granularity task-level planning needs: "will swinging here produce a log" is not
answerable from a next-frame latent.

Measured on 2x RTX 6000 Ada (details in docs/stage1b-log.md):

  event prediction   R@5 0.764, mAP 0.647  vs persistence baseline mAP 0.305
                     positive-slot p=0.276 vs negative-slot p=0.00036

  The training BCE falls to 0.001, which is meaningless -- 1024-way multi-label
  with 1.57 positives per step is won by predicting all zeros. The numbers above
  are what the head actually learned.

Also wired EventEmbedding.init_from_text, which had no call site anywhere in the
repo (getting-started.md records the gap). Concept rows were random, so
"mined oak log" and "mined dark oak log" were unrelated. They are now seeded by
running each phrase through the backbone and taking hidden states -- not by
looking up the embedding table the way GoalEncoder.from_text does.

Negative results, all in the log: a 3x larger backbone changes nothing
(1.2559 vs 1.2266); eight random vectors match the text embeddings (1.2501 vs
1.2266); unseen paraphrases transfer at 18% against a 12.5% chance floor. The
common cause is that from_text only touches the input embedding layer, so the
LLM's 28 layers never encode an instruction at all.

Fixes: eval_stage1a.py hit the documented bfloat16-to-numpy trap that minerl.py
and play_server.py already guard against; play_server streamed MJPEG over
HTTP/1.0 with no per-part Content-Length, which buffering proxies swallow whole,
so the page froze on its first frame -- it now polls /frame.jpg instead.

+++ wam/training/eval_stage1a.py
@@ -85,8 +85,12 @@ def run_instruction(model, env, instruction: str, goal_table, steps: int = 20):
         # Bin indices, not degrees: the autocorrelation is about whether
         # successive decisions are related, and bins keep that while being
         # immune to the mu-law's nonlinearity.
-        cameras.append(action.camera[0].cpu().numpy())
-        pressed += action.buttons[0].cpu().numpy().mean(axis=0)
+        # .float() 不能省：ActionChunk 带着主干的 dtype，模型跑在 bfloat16 上，
+        # 而 numpy 没有这个 dtype。手工构造的测试 chunk 是 float32，所以这行能
+        # 通过任何 dry run，只在真模型驱动时才炸 —— measurements.md 记过同一个坑，
+        # minerl.py 和 play_server.py 都防住了，唯独这里漏了。
+        cameras.append(action.camera[0].long().cpu().numpy())
+        pressed += action.buttons[0].float().cpu().numpy().mean(axis=0)
 
         obs = env.step(action)
         if obs.event_text and "mined" in obs.event_text:
